+++ mitra_activitypub/src/lib.rs
@@ -3,6 +3,7 @@ pub mod agent;
 pub mod authentication;
 pub mod authority;
 pub mod builders;
+mod c2s;
 mod constants;
 mod contexts;
 mod deliverer;